+++ survwey/survwey/src/main/java/mcnc/survwey/domain/survey/common/dto/SurveyDTO.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
+import mcnc.survwey.domain.survey.common.Survey;
 
 
 import java.time.LocalDateTime;
@@ -22,4 +23,13 @@ public class SurveyDTO {
     @NotNull(message = "만료일 지정은 필수입니다.")
     private LocalDateTime expireDate;
 
+    public static SurveyDTO of(Survey survey) {
+        return SurveyDTO.builder()
+                .surveyId(survey.getSurveyId())
+                .title(survey.getTitle())
+                .description(survey.getDescription())
+                .createDate(survey.getCreateDate())
+                .expireDate(survey.getExpireDate())
+                .build();
+    }
 }
